CMS/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
Django==4.2
django-oauth-toolkit==2.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.4.2
oauthlib==3.2.2
Pillow==9.5.0
pycparser==2.21
pytz==2023.3
requests==2.28.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
